+++ test/unit/code_test.rb
@@ -1,15 +1,9 @@
 require 'minitest/autorun'
 require 'dtext'
+require_relative 'test_helper'
 
 class DTextCodeTest < Minitest::Test
-  def assert_parse(expected, input, **options)
-    if expected.nil?
-      assert_nil(str_part = DText.parse(input, **options))
-    else
-      str_part = DText.parse(input, **options)[0]
-      assert_equal(expected, str_part)
-    end
-  end
+  include DTextTestHelper
 
   def test_inline_code_backticks
     assert_parse(%{<p><span class="inline-code">`what`</span></p>}, "`\\`what\\``")
